Add functionality to index.js
